clean up

+++ src/main/java/sk/stopangin/spring/l1/ExpenseService.java
@@ -2,6 +2,6 @@ package sk.stopangin.spring.l1;
 
 public interface ExpenseService {
 
-    Integer getSumOfAllExpenses();
+    double getSumOfAllExpenses();
 
 }
